+++ app/schemas/qr_token.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 class QrTokenGenerate(BaseModel):
     cesto_id: str
     cantidad_botellas: int
-    peso: float
+    peso: int
     is_used: bool = False
